+++ blocklets/ai-kit/src/libs/app.ts
@@ -1,4 +1,4 @@
-import { TSubscriptionExpanded } from '@did-pay/client';
+import { TSubscriptionExpanded } from '@blocklet/payment-js';
 
 import api from './api';