+++ blocklets/ai-kit/src/locales/en.ts
@@ -12,4 +12,10 @@ export default flat({
   unsubscribeTip: 'After unsubscribing, you will no longer be able to continue using the AI services we provide!',
   cancel: 'Cancel',
   unsubscribeAt: 'Unsubscribe at',
+  cancelled: 'Cancelled',
+  recoverSubscription: 'Recover Subscription',
+  recoverSubscriptionTip: 'After recover the subscription, you can use the AI services we provide!',
+  recoverSubscriptionSucceed: 'Subscription recover successful!',
+  total: 'Total',
+  monthlySpend: 'Monthly Spend',
 });